Mysql Dump before shutdown/reboot

+++ project/manifests/web.pp
@@ -7,6 +7,7 @@ class project::web($phpmyadmin = true){
 
 		exec{ "link phpmyadmin config":
 			command => "ln -s /etc/phpmyadmin/apache.conf /etc/apache2/conf.d/phpmyadmin.conf",
+			unless => "/bin/readlink -e /etc/apache2/conf.d/phpmyadmin.conf",
 			require => Package["phpmyadmin"],
 			notify => Service["apache2"],
 		}
